fixed conflicts
