+++ collect_app/src/main/java/org/odk/collect/android/activities/FormEntryActivity.java
@@ -15,6 +15,7 @@
 package org.odk.collect.android.activities;
 
 import android.app.DatePickerDialog;
+import android.app.TimePickerDialog;
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
 import android.content.Context;
@@ -47,6 +48,7 @@ import android.widget.EditText;
 import android.widget.FrameLayout;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+import android.widget.TimePicker;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -160,6 +162,7 @@ import org.odk.collect.android.widgets.utilities.FormControllerWaitingForDataReg
 import org.odk.collect.android.widgets.utilities.WaitingForDataRegistry;
 import org.odk.collect.async.Scheduler;
 import org.odk.collect.audioclips.AudioClipViewModel;
+import org.odk.collect.android.widgets.TimeWidget;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -214,7 +217,9 @@ public class FormEntryActivity extends CollectAbstractActivity implements Animat
         ScreenContext, FormLoadingDialogFragment.FormLoadingDialogFragmentListener,
         AudioControllerView.SwipableParent, FormIndexAnimationHandler.Listener,
         QuitFormDialogFragment.Listener, DeleteRepeatDialogFragment.DeleteRepeatDialogCallback,
-        SelectMinimalDialog.SelectMinimalDialogListener, DatePickerDialog.OnDateSetListener {
+        SelectMinimalDialog.SelectMinimalDialogListener, DatePickerDialog.OnDateSetListener,
+        TimePickerDialog.OnTimeSetListener {
+
 
     // Defines for FormEntryActivity
     private static final boolean EXIT = true;
@@ -722,6 +727,19 @@ public class FormEntryActivity extends CollectAbstractActivity implements Animat
                             .withSecondOfMinute(0)
                             .withMillisOfSecond(0);
                     ((DateWidget) qw).setBinaryData(date);
+                    widgetValueChanged(qw);
+                    return;
+                }
+            }
+        }
+    }
+
+    @Override
+    public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
+        if (currentView != null) {
+            for (QuestionWidget qw : ((ODKView) currentView).getWidgets()) {
+                if (qw instanceof TimeWidget) {
+                    ((TimeWidget) qw).onTimeSet(view, hourOfDay, minute);
                     widgetValueChanged(qw);
                     return;
                 }
