+++ collect_app/src/main/java/org/odk/collect/android/widgets/TimeWidget.java
@@ -15,15 +15,8 @@
 package org.odk.collect.android.widgets;
 
 import android.annotation.SuppressLint;
-import android.app.TimePickerDialog;
 import android.content.Context;
-import android.content.res.TypedArray;
-import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
-import android.os.Build;
-import android.text.format.DateFormat;
-import android.util.AttributeSet;
-import android.view.Window;
+import android.os.Bundle;
 import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.TextView;
@@ -33,28 +26,21 @@ import org.javarosa.core.model.data.IAnswerData;
 import org.javarosa.core.model.data.TimeData;
 import org.joda.time.DateTime;
 import org.odk.collect.android.R;
+import org.odk.collect.android.activities.FormEntryActivity;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
 import org.odk.collect.android.formentry.questions.WidgetViewUtils;
 import org.odk.collect.android.widgets.interfaces.ButtonClickListener;
-
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
+import org.odk.collect.android.fragments.dialogs.CustomTimePickerDialog;
+import org.odk.collect.android.utilities.DialogUtils;
 import java.util.Date;
-
-import timber.log.Timber;
 
 import static org.odk.collect.android.formentry.questions.WidgetViewUtils.createAnswerTextView;
 import static org.odk.collect.android.formentry.questions.WidgetViewUtils.createSimpleButton;
+import static org.odk.collect.android.fragments.dialogs.CustomTimePickerDialog.CURRENT_TIME;
+import static org.odk.collect.android.fragments.dialogs.CustomTimePickerDialog.TIME_PICKER_THEME;
 
-/**
- * Displays a TimePicker widget.
- *
- * @author Carl Hartung (carlhartung@gmail.com)
- */
 @SuppressLint("ViewConstructor")
-public class TimeWidget extends QuestionWidget implements ButtonClickListener, TimePickerDialog.OnTimeSetListener {
-    private TimePickerDialog timePickerDialog;
-
+public class TimeWidget extends QuestionWidget implements ButtonClickListener {
     Button timeButton;
     final TextView timeTextView;
 
@@ -71,7 +57,15 @@ public class TimeWidget extends QuestionWidget implements ButtonClickListener, T
         super(context, prompt, !isPartOfDateTimeWidget);
         createTimeButton();
         timeTextView = createAnswerTextView(getContext(), getAnswerFontSize());
-        createTimePickerDialog();
+
+        if (getFormEntryPrompt().getAnswerValue() == null) {
+            clearAnswer();
+        } else {
+            Date date = (Date) getFormEntryPrompt().getAnswerValue().getValue();
+
+            DateTime dateTime = new DateTime(date);
+            updateTime(dateTime, true);
+        }
         addViews();
     }
 
@@ -128,17 +122,11 @@ public class TimeWidget extends QuestionWidget implements ButtonClickListener, T
     }
 
     private void createTimePickerDialog() {
-        timePickerDialog = new CustomTimePickerDialog(getContext(), this, 0, 0);
-        timePickerDialog.setCanceledOnTouchOutside(false);
+        Bundle bundle = new Bundle();
+        bundle.putInt(TIME_PICKER_THEME, themeUtils.getHoloDialogTheme());
+        bundle.putSerializable(CURRENT_TIME, new DateTime().withTime(hourOfDay, minuteOfHour, 0, 0));
 
-        if (getFormEntryPrompt().getAnswerValue() == null) {
-            clearAnswer();
-        } else {
-            Date date = (Date) getFormEntryPrompt().getAnswerValue().getValue();
-
-            DateTime dateTime = new DateTime(date);
-            updateTime(dateTime, true);
-        }
+        DialogUtils.showIfNotShowing(CustomTimePickerDialog.class, bundle, ((FormEntryActivity) getContext()).getSupportFragmentManager());
     }
 
     public int getHour() {
@@ -157,10 +145,6 @@ public class TimeWidget extends QuestionWidget implements ButtonClickListener, T
         updateTime(DateTime.now(), false);
     }
 
-    public void updateTime(DateTime dateTime) {
-        updateTime(dateTime, true);
-    }
-
     public void updateTime(DateTime dateTime, boolean shouldUpdateLabel) {
         updateTime(dateTime.getHourOfDay(), dateTime.getMinuteOfHour(), shouldUpdateLabel);
     }
@@ -169,22 +153,16 @@ public class TimeWidget extends QuestionWidget implements ButtonClickListener, T
         this.hourOfDay = hourOfDay;
         this.minuteOfHour = minuteOfHour;
 
-        timePickerDialog.updateTime(hourOfDay, minuteOfHour);
-
         if (shouldUpdateLabel) {
             setTimeLabel();
         }
     }
 
-    @Override
     public void onTimeSet(TimePicker timePicker, int hourOfDay, int minute) {
         timePicker.clearFocus();
-
-        this.hourOfDay = timePicker.getCurrentHour();
-        this.minuteOfHour = timePicker.getCurrentMinute();
-
+        this.hourOfDay = hourOfDay;
+        this.minuteOfHour = minute;
         setTimeLabel();
-        widgetValueChanged();
     }
 
     @Override
@@ -194,116 +172,6 @@ public class TimeWidget extends QuestionWidget implements ButtonClickListener, T
         } else {
             updateTime(hourOfDay, minuteOfHour, true);
         }
-        timePickerDialog.show();
-    }
-
-    private class CustomTimePickerDialog extends TimePickerDialog {
-        private final String dialogTitle = getContext().getString(R.string.select_time);
-
-        CustomTimePickerDialog(Context context, OnTimeSetListener callBack, int hour, int minute) {
-            super(context, themeUtils.getHoloDialogTheme(), callBack, hour, minute, DateFormat.is24HourFormat(context));
-            setTitle(dialogTitle);
-            fixSpinner(context, hour, minute, DateFormat.is24HourFormat(context));
-
-            Window window = getWindow();
-            if (window != null) {
-                window.setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
-            }
-        }
-
-        public void setTitle(CharSequence title) {
-            super.setTitle(dialogTitle);
-        }
-
-        /**
-         * Workaround for this bug: https://code.google.com/p/android/issues/detail?id=222208
-         * In Android 7.0 Nougat, spinner mode for the TimePicker in TimePickerDialog is
-         * incorrectly displayed as clock, even when the theme specifies otherwise.
-         * <p>
-         * Source: https://gist.github.com/jeffdgr8/6bc5f990bf0c13a7334ce385d482af9f
-         */
-        @SuppressWarnings("deprecation")
-        private void fixSpinner(Context context, int hourOfDay, int minute, boolean is24HourView) {
-            if (Build.VERSION.SDK_INT == Build.VERSION_CODES.N) {
-                try {
-                    // Get the theme's android:timePickerMode
-                    final int MODE_SPINNER = 2;
-                    Class<?> styleableClass = Class.forName("com.android.internal.R$styleable");
-                    Field timePickerStyleableField = styleableClass.getField("TimePicker");
-                    int[] timePickerStyleable = (int[]) timePickerStyleableField.get(null);
-                    final TypedArray a = context.obtainStyledAttributes(null, timePickerStyleable,
-                            android.R.attr.timePickerStyle, 0);
-                    Field timePickerModeStyleableField = styleableClass.getField("TimePicker_timePickerMode");
-                    int timePickerModeStyleable = timePickerModeStyleableField.getInt(null);
-                    final int mode = a.getInt(timePickerModeStyleable, MODE_SPINNER);
-                    a.recycle();
-
-                    if (mode == MODE_SPINNER) {
-                        Field field = findField(TimePickerDialog.class, TimePicker.class, "mTimePicker");
-                        if (field == null) {
-                            Timber.e("Reflection failed: Couldn't find field 'mTimePicker'");
-                            return;
-                        }
-
-                        TimePicker timePicker = (TimePicker) field.get(this);
-                        Class<?> delegateClass = Class.forName("android.widget.TimePicker$TimePickerDelegate");
-                        Field delegateField = findField(TimePicker.class, delegateClass, "mDelegate");
-
-                        if (delegateField == null) {
-                            Timber.e("Reflection failed: Couldn't find field 'mDelegate'");
-                            return;
-                        }
-                        Object delegate = delegateField.get(timePicker);
-
-                        Class<?> spinnerDelegateClass;
-                        spinnerDelegateClass = Class.forName("android.widget.TimePickerSpinnerDelegate");
-                        // In 7.0 Nougat for some reason the timePickerMode is ignored and the
-                        // delegate is TimePickerClockDelegate
-                        if (delegate.getClass() != spinnerDelegateClass) {
-                            delegateField.set(timePicker, null); // throw out the TimePickerClockDelegate!
-                            timePicker.removeAllViews(); // remove the TimePickerClockDelegate views
-                            Constructor spinnerDelegateConstructor = spinnerDelegateClass
-                                    .getConstructor(TimePicker.class, Context.class,
-                                            AttributeSet.class, int.class, int.class);
-                            spinnerDelegateConstructor.setAccessible(true);
-
-                            // Instantiate a TimePickerSpinnerDelegate
-                            delegate = spinnerDelegateConstructor.newInstance(timePicker, context,
-                                    null, android.R.attr.timePickerStyle, 0);
-
-                            // set the TimePicker.mDelegate to the spinner delegate
-                            delegateField.set(timePicker, delegate);
-
-                            // Set up the TimePicker again, with the TimePickerSpinnerDelegate
-                            timePicker.setIs24HourView(is24HourView);
-                            timePicker.setCurrentHour(hourOfDay);
-                            timePicker.setCurrentMinute(minute);
-                            timePicker.setOnTimeChangedListener(this);
-                        }
-                    }
-                } catch (Exception e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-
-        private Field findField(Class objectClass, Class fieldClass, String expectedName) {
-            try {
-                Field field = objectClass.getDeclaredField(expectedName);
-                field.setAccessible(true);
-                return field;
-            } catch (NoSuchFieldException e) {
-                Timber.i(e); // ignore
-            }
-
-            // search for it if it wasn't found under the expected ivar name
-            for (Field searchField : objectClass.getDeclaredFields()) {
-                if (searchField.getType() == fieldClass) {
-                    searchField.setAccessible(true);
-                    return searchField;
-                }
-            }
-            return null;
-        }
+        createTimePickerDialog();
     }
 }
